Creating base class BaseRackDeviceExtractor
