+++ .vscode/launch.json
@@ -11,7 +11,7 @@
             "args": [
                 "initdb",
                 "--files",
-                "/Users/zenluiz/Desktop/Testes ALS/Projects folder 1/MidiFxRack Project/MidiFxRack.als",
+                "/Users/zenluiz/Desktop/Testes ALS/Projects folder 1/DrumRacks Project/DrumRacks.als",
                 "--log-level",
                 "verbose"
             ],
