+++ src/als-tools.infrastructure/Extractors/DeviceTypes/StockDevices/StockRacks/IStockRackDeviceExtractor.cs
@@ -2,4 +2,5 @@ namespace AlsTools.Infrastructure.Extractors.DeviceTypes.StockDevices.StockRacks
 
 public interface IStockRackDeviceExtractor : IStockDeviceExtractor
 {
+    string XPathDevicesSelector { get; }
 }
